+++ tests/App/Presenters/UserPresenter.php
@@ -47,6 +47,8 @@ class UserPresenter extends Presenter implements Searchable, Personable
      */
     public function image(): ?string
     {
-        return $this->entity->getAvatar();
+        $hash = md5(strtolower(trim($this->entity->email)));
+
+        return "https://www.gravatar.com/avatar/$hash";
     }
 }