+++ src/Screen/Contracts/Searchable.php
@@ -4,8 +4,24 @@ declare(strict_types=1);
 
 namespace Orchid\Screen\Contracts;
 
+use Laravel\Scout\Builder;
+
 interface Searchable
 {
+    /**
+     * The number of models to return for show compact search result.
+     *
+     * @return int
+     */
+    public function perSearchShow(): int;
+
+    /**
+     * @param string $query
+     *
+     * @return Builder
+     */
+    public function searchQuery(string $query = null): Builder;
+
     /**
      * @return string
      */