+++ tests/App/Presenters/UserPresenter.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Orchid\Tests\App\Presenters;
 
+use Laravel\Scout\Builder;
 use Orchid\Screen\Contracts\Personable;
 use Orchid\Screen\Contracts\Searchable;
 use Orchid\Support\Presenter;
@@ -50,5 +51,21 @@ class UserPresenter extends Presenter implements Searchable, Personable
         $hash = md5(strtolower(trim($this->entity->email)));
 
         return "https://www.gravatar.com/avatar/$hash";
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function perSearchShow(): int
+    {
+        return 3;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function searchQuery(string $query = null): Builder
+    {
+        return $this->entity->search($query);
     }
 }